refactor(ui): extract ChatView scroll lifecycle
